Add note on which options must only be set using command line (see details).
As #229 and #297 in -core state, people try to set every possible option
in init file. On server start and pid file drop, though, init file(s) are
not yet loaded. So, setting pid file or port in init file makes no sense.
Defaults are being used and it causes confusion in future.

Given all this, add a note in generated init file.